+++ src/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../App.css';
-import BooksForm from './BooksForm';
-import BooksList from './BooksList';
+import BooksForm from '../containers/BooksForm';
+import BooksList from '../containers/BooksList';
 
 const App = () => (
   <div>
