+++ src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import App from './components/App';
-import booksReducer from './reducers/books';
+import booksReducer from './reducers/index';
 
 const defaultState = {
   firstBook: {
